+++ word/src/main/java/com/dreampany/word/data/source/repository/WordRepository.java
@@ -229,12 +229,12 @@ public class WordRepository extends Repository<String, Word> implements WordData
     }
 
     @Override
-    public List<Word> getItems(long limit) {
+    public List<Word> getItems(int limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Word>> getItemsRx(long limit) {
+    public Maybe<List<Word>> getItemsRx(int limit) {
         return null;
     }
