+++ word/src/main/java/com/dreampany/word/data/source/room/WordRoomDataSource.java
@@ -161,12 +161,12 @@ public class WordRoomDataSource implements WordDataSource {
     }
 
     @Override
-    public List<Word> getItems(long limit) {
+    public List<Word> getItems(int limit) {
         return null;
     }
 
     @Override
-    public Maybe<List<Word>> getItemsRx(long limit) {
+    public Maybe<List<Word>> getItemsRx(int limit) {
         return null;
     }
